requirements.txt
streamlit==1.38.0
tensorflow==2.13.0
Pillow==9.4.0
numpy==1.24.3
requests==2.31.0
